+++ configure.ac
@@ -575,6 +575,7 @@ src/libs/Makefile
 src/libs/zmbv/Makefile
 src/libs/gui_tk/Makefile
 src/libs/decoders/Makefile
+src/libs/nuked/Makefile
 src/misc/Makefile
 src/shell/Makefile
 src/platform/Makefile
