+++ src/Makefile.am
@@ -14,6 +14,9 @@ endif
 dosbox_SOURCES = dosbox.cpp $(ico_stuff)
 dosbox_LDADD = cpu/libcpu.a debug/libdebug.a dos/libdos.a fpu/libfpu.a  hardware/libhardware.a gui/libgui.a \
                ints/libints.a misc/libmisc.a shell/libshell.a hardware/mame/libmame.a \
-               hardware/serialport/libserial.a libs/gui_tk/libgui_tk.a libs/decoders/libdecoders.a
+               hardware/serialport/libserial.a \
+               libs/decoders/libdecoders.a \
+               libs/gui_tk/libgui_tk.a \
+               libs/nuked/libnuked.a
 
 EXTRA_DIST = winres.rc dosbox.ico
