+++ src/libs/nuked/Makefile.am
@@ -1,0 +1,7 @@
+AM_CPPFLAGS = -I$(top_srcdir)/include
+
+EXTRA_DIST = nukedopl.h
+
+noinst_LIBRARIES = libnuked.a
+
+libnuked_a_SOURCES = nukedopl.cpp
